pyproject.toml
[project]
name = "aecl-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.4",
]

[project.scripts]
aecl_bot=".__main__:main"